+++ data/tasks.txt
@@ -1,1 +1,2 @@
 E///nightrun///2098-09-09///false
+T///skydive/// ///false
