+++ src/main/java/unc/UncException.java
@@ -1,3 +1,5 @@
+package unc;
+
 public class UncException extends Exception{
     public UncException(String message) {
         super(message);
