+++ src/main/java/unc/command/DeleteCommand.java
@@ -1,0 +1,27 @@
+package unc.command;
+
+import unc.Storage;
+import unc.task.Task;
+import unc.TaskList;
+import unc.Ui;
+
+public class DeleteCommand extends Command {
+    private final int index;
+
+    public DeleteCommand(int num) {
+        this.index = num - 1;
+    }
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        Task tempTask = taskList.get(index);
+        taskList.delete(index);
+        ui.delete(tempTask, taskList.size());
+        storage.save(taskList);
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
