+++ src/main/java/unc/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package unc.command;
+
+import unc.Storage;
+import unc.TaskList;
+import unc.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
