+++ src/main/java/unc/command/UnmarkCommand.java
@@ -1,15 +1,20 @@
-public class DeleteCommand extends Command {
+package unc.command;
+
+import unc.Storage;
+import unc.TaskList;
+import unc.Ui;
+
+public class UnmarkCommand extends Command {
     private final int index;
 
-    public DeleteCommand(int num) {
+    public UnmarkCommand(int num) {
         this.index = num - 1;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        Task tempTask = taskList.get(index);
-        taskList.delete(index);
-        ui.delete(tempTask, taskList.size());
+        taskList.markAsNotDone(index);
+        ui.unmark(taskList, index);
         storage.save(taskList);
     }
