+++ src/main/java/unc/task/Event.java
@@ -1,24 +1,28 @@
+package unc.task;
+
+import unc.UncException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task {
+public class Event extends Task {
 
-    protected LocalDate by;
+    protected LocalDate at;
 
-    public Deadline(String description, String by) throws UncException {
+    public Event(String description, String at) throws UncException {
         super(description);
         try {
-            this.by = LocalDate.parse(by);
+            this.at = LocalDate.parse(at);
         } catch (DateTimeParseException exception) {
             throw new UncException("Invalid datetime");
         }
     }
 
-    public Deadline(String description, String by, String done) throws UncException {
+    public Event(String description, String at, String done) throws UncException {
         super(description, done == "true");
         try {
-            this.by = LocalDate.parse(by);
+            this.at = LocalDate.parse(at);
         } catch (DateTimeParseException exception) {
             throw new UncException("Invalid datetime");
         }
@@ -26,11 +30,12 @@ public class Deadline extends Task {
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("dd-MM-yyyy")) + ")";
+        return "[E]" + super.toString() + " (at: " + at.format(DateTimeFormatter.ofPattern("dd-MM-yyyy")) + ")";
     }
+
 
     @Override
     public String toStorageString() {
-        return "D" + "///" + this.description + "///" + this.by + "///" + this.isDone;
+        return "E" + "///" + this.description + "///" + this.at + "///" + this.isDone;
     }
 }
